requirements.txt
os
pandas
streamlit
google-generativeai